Fix typing issues
