+++ sail_on_client/protocol/ond_dataclasses.py
@@ -15,7 +15,7 @@ from sail_on_client.feedback.activity_recognition_feedback import (
 )
 
 try:
-    from importlib.metadata import version, PackageNotFoundError
+    from importlib.metadata import version, PackageNotFoundError  # type:ignore
 except ModuleNotFoundError:
     from importlib_metadata import version, PackageNotFoundError
