+++ sail_on_client/protocol/ond_protocol.py
@@ -54,7 +54,7 @@ class SailOn(VisualProtocol):
         baseline_algorithm_name: str,
         has_baseline: bool,
         has_reaction_baseline: bool,
-        test_ids,
+        test_ids: List[str],
     ) -> AlgorithmAttributes:
         """
         Create an instance of algorithm attributes.
@@ -84,6 +84,49 @@ class SailOn(VisualProtocol):
             session_id,
             test_ids,
         )
+
+    def create_algorithm_session(
+        self,
+        algorithm_attributes: AlgorithmAttributes,
+        domain: str,
+        hints: List[str],
+        has_a_session: bool,
+        protocol_name: str,
+    ) -> AlgorithmAttributes:
+        """
+        Create/resume session for an algorithm.
+
+        Args:
+            algorithm_attributes: An instance of AlgorithmAttributes
+            domain: Domain for the algorithm
+            hints: List of hints used in the session
+            has_a_session: Already has a session and we want to resume it
+            protocol_name: Name of the algorithm
+
+        Returns:
+            An AlgorithmAttributes object with updated session id or test id
+        """
+        test_ids = algorithm_attributes.test_ids
+        named_version = algorithm_attributes.named_version()
+        detection_threshold = algorithm_attributes.detection_threshold
+
+        if has_a_session:
+            session_id = algorithm_attributes.session_id
+            finished_test = self.harness.resume_session(session_id)
+            algorithm_attributes.remove_completed_tests(finished_test)
+            log.info(f"Resumed session {session_id} for {algorithm_attributes.name}")
+        else:
+            session_id = self.harness.session_request(
+                test_ids,
+                protocol_name,
+                domain,
+                named_version,
+                hints,
+                detection_threshold,
+            )
+            algorithm_attributes.session_id = session_id
+            log.info(f"Created session {session_id} for {algorithm_attributes.name}")
+        return algorithm_attributes
 
     def _find_baseline_session_id(
         self, algorithms_attributes: List[AlgorithmAttributes]
